refactor(core): move the exec-tool registry from scripts/lib into src/core/exec-tools
scripts/lib/registry.ts is the production exec-tool registry (globalThis
Map + PluginToolContext builder), not build tooling — moved to
src/core/exec-tools/registry.ts. Its 5 production importers (mcp-server,
bakin-skill, reload-pipeline, plugin-registry, the exec-tools API route),
the 5 scripts/lib self-registering tool peers, and all test importers are
repointed; the dedicated test moves to tests/core/exec-tools/. Combined
with K1, this collapses the dependency cycle from 18 → 6: the entire
scripts/lib/registry cluster (13 cycles) is gone. Relocating the 5 tool
peers out of scripts/ entirely is a noted WS7-tooling follow-up. No
behavior change (exec-tool registration unchanged — mcp-server tests green).
