+++ scripts/lib/heartbeat.ts
@@ -7,7 +7,7 @@ import { writeFileSync, mkdirSync, existsSync } from 'fs'
 import { join } from 'path'
 import { getContentDir } from '../../src/core/content-dir'
 import { recordUsage } from '../../src/core/usage'
-import { addExecTool } from './registry'
+import { addExecTool } from '../../src/core/exec-tools/registry'
 
 addExecTool({
   name: 'bakin_exec_heartbeat',
