+++ scripts/lib/log-progress.ts
@@ -4,7 +4,7 @@
 import { z } from 'zod'
 import { logProgress } from '../../src/core/task-service'
 import { succeed, fail } from './common'
-import { addExecTool } from './registry'
+import { addExecTool } from '../../src/core/exec-tools/registry'
 import type { ExecToolResult } from '@bakin/core/plugin-types'
 
 const CATEGORIES = ['start', 'progress', 'milestone', 'blocked', 'complete'] as const
