+++ src/core/plugin-host/reload-pipeline.ts
@@ -37,7 +37,7 @@ import { join } from 'path'
 import { createLogger } from '@/core/logger'
 import { pluginRegistry, removePluginSkillsByPlugin } from '@/lib/plugin-registry'
 import { getHookRegistry } from '@bakin/core/hooks/hook-registry-singleton'
-import { removeExecToolsByPlugin } from '../../../scripts/lib/registry'
+import { removeExecToolsByPlugin } from '../exec-tools/registry'
 import { unregisterPluginNodeTypes } from '../../../plugins/workflows/lib/node-type-registry'
 import { unregisterPluginNotificationChannels } from '../../../plugins/workflows/lib/notification-channel-registry'
 import { unregisterPluginHealthChecks } from '../health-check-registry'
