+++ tests/integration/usage-wiring-mcp.test.ts
@@ -87,7 +87,7 @@ describe('MCP usage wiring (integration)', () => {
   })
 
   it('records a successful tool invocation as an mcp usage entry', async () => {
-    const { addExecTool } = require('../../scripts/lib/registry') as typeof import('../../scripts/lib/registry')
+    const { addExecTool } = require('@/core/exec-tools/registry') as typeof import('@/core/exec-tools/registry')
     const { registerTools } = require('../../src/core/mcp-server') as typeof import('../../src/core/mcp-server')
     const { getUsageFeed } = require('../../src/core/usage') as typeof import('../../src/core/usage')
 
@@ -120,7 +120,7 @@ describe('MCP usage wiring (integration)', () => {
   })
 
   it('records a not-ok handler result as an error entry', async () => {
-    const { addExecTool } = require('../../scripts/lib/registry') as typeof import('../../scripts/lib/registry')
+    const { addExecTool } = require('@/core/exec-tools/registry') as typeof import('@/core/exec-tools/registry')
     const { registerTools } = require('../../src/core/mcp-server') as typeof import('../../src/core/mcp-server')
     const { getUsageFeed } = require('../../src/core/usage') as typeof import('../../src/core/usage')
 
@@ -146,7 +146,7 @@ describe('MCP usage wiring (integration)', () => {
   })
 
   it('records a thrown handler exception as an error entry', async () => {
-    const { addExecTool } = require('../../scripts/lib/registry') as typeof import('../../scripts/lib/registry')
+    const { addExecTool } = require('@/core/exec-tools/registry') as typeof import('@/core/exec-tools/registry')
     const { registerTools } = require('../../src/core/mcp-server') as typeof import('../../src/core/mcp-server')
     const { getUsageFeed } = require('../../src/core/usage') as typeof import('../../src/core/usage')
 
@@ -175,7 +175,7 @@ describe('MCP usage wiring (integration)', () => {
 
   it('regression sentinel: three invocations produce count=3, errors=2', async () => {
     const { clearUsage, getUsageFeed } = require('../../src/core/usage') as typeof import('../../src/core/usage')
-    const { addExecTool } = require('../../scripts/lib/registry') as typeof import('../../scripts/lib/registry')
+    const { addExecTool } = require('@/core/exec-tools/registry') as typeof import('@/core/exec-tools/registry')
     const { registerTools } = require('../../src/core/mcp-server') as typeof import('../../src/core/mcp-server')
 
     clearUsage()
